requirements.txt
tensorflow>=2.5.0
opencv-python>=4.5.3
numpy>=1.19.5
pillow>=8.3.1
pytesseract>=0.3.8
scikit-image>=0.18.3